third_party/WebKit/LayoutTests/platform/win/editing/inserting/insert_html_as_plain_text-expected.txt

This is a testharness.js-based test.
PASS insert HTML as plain text. 
PASS insert HTML at start of paragraph. 
PASS interchange-newline is not inserted before DIV 
Harness: the test ran to completion.
